+++ backend/canisters/group/impl/src/updates/unfollow_thread.rs
@@ -20,6 +20,6 @@ fn unfollow_thread_impl(args: Args, state: &mut RuntimeState) -> OCResult {
         .chat
         .unfollow_thread(user_id, args.thread_root_message_index, now)?;
 
-    state.data.mark_group_updated_in_user_canister(user_id);
+    state.mark_activity_for_user(user_id);
     Ok(())
 }