+++ backend/libraries/group_chat_core/src/members.rs
@@ -630,6 +630,7 @@ impl GroupMemberInternal {
             self.suspended.timestamp,
             self.rules_accepted.as_ref().map(|r| r.timestamp).unwrap_or_default(),
             self.lapsed.timestamp,
+            self.proposal_votes.last().map(|(ts, _)| *ts).unwrap_or_default(),
         ]
         .into_iter()
         .max()