+++ backend/canisters/community/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Remove duplication by removing `MessageContent::message_type` function ([#8293](https://github.com/open-chat-labs/open-chat/pull/8293))
+
 ### Removed
 
 - Removed unused fields from BotChatEvent ([#8291](https://github.com/open-chat-labs/open-chat/pull/8291))