+++ backend/canisters/community/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.1829](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1829-community)] - 2025-07-09
+
 ### Added
 
 - Introduce `Encrypted` message type ([8294](https://github.com/open-chat-labs/open-chat/pull/8294))